Refactor schemas out of resources (to avoid deps, circular import hell)
