+++ neuroscout/requirements.txt
@@ -1,10 +1,34 @@
 apispec==1.1.0
-appdirs==1.4.3
-blinker==1.4
 citeproc-py==0.4.0
 citeproc-py-styles==0.1.1
-click==6.7
 celery
+Flask==0.12.4
+flask-apispec==0.8.0
+Flask-cors==3.0.2
+Flask-Caching==1.3.3
+Flask-JWT==0.3.2
+Flask-Mail==0.9.1
+Flask-Migrate==2.0.3
+Flask-Principal==0.4.0
+Flask-Security==1.7.5
+Flask-Script==2.0.5
+Flask-SQLAlchemy==2.2
+Flask-WTF==0.14.2
+gunicorn==19.7.1
+itsdangerous==0.24
+MarkupSafe==1.0
+marshmallow==2.19.1
+psycopg2==2.7.1
+PyJWT==1.4.2
+pyparsing==2.2.0
+redis
+pandas==0.23
+nibabel==2.1.0
+scipy
+patsy
+pybids==0.8.0
+appdirs==1.4.3
+blinker==1.4
 decorator==4.0.11
 pyld==0.7.3
 datalad==0.10.2
@@ -23,44 +47,32 @@ Flask-Security==1.7.5
 Flask-Script==2.0.5
 Flask-SQLAlchemy==2.2
 Flask-WTF==0.14.2
-google-api-python-client
 gunicorn==19.7.1
+google-api-python-client
 hashids==1.2.0
-imageio==2.5.0
 ipython==6.2.1
-itsdangerous==0.24
-imageio-ffmpeg
 Jinja2==2.10.1
-MarkupSafe==1.0
-marshmallow==2.19.1
-moviepy==0.2.3.5
-nibabel==2.1.0
 nltk==3.2.2
 numpy==1.12.0
 olefile==0.44
-opencv-python
 packaging==16.8
-pandas==0.22
-patsy
 passlib==1.7.1
 pytesseract
 Pillow==5.4.1
-psycopg2==2.7.1
 progressbar2
 py==1.4.33
-pybids==0.8.0
-PyJWT==1.4.2
-pyparsing==2.2.0
 pytest==3.0.7
 python-dateutil==2.6.0
 python-magic==0.4.13
 pytz==2017.2
-redis
 requests==2.21.0
-scipy
 SQLAlchemy==1.3.0
 tqdm==4.11.2
 webargs>=5.1.3
 Werkzeug==0.12.1
 WTForms==2.1
 xlrd
+moviepy==0.2.3.5
+opencv-python
+imageio-ffmpeg
+imageio==2.5.0
