+++ neuroscout/resources/task.py
@@ -1,17 +1,8 @@
-from marshmallow import Schema, fields
 from flask_apispec import MethodResource, marshal_with, doc, use_kwargs
 from ..models import Task
 import webargs as wa
 from .utils import first_or_404
-
-class TaskSchema(Schema):
-    id = fields.Int()
-    name = fields.Str()
-
-    dataset_id = fields.Int()
-    num_runs = fields.Int(description='Number of runs.')
-    TR = fields.Number()
-    summary = fields.Str(description='Task summary description')
+from ..schemas.task import TaskSchema
 
 
 class TaskResource(MethodResource):
