+++ neuroscout/resources/user.py
@@ -1,54 +1,13 @@
 from flask_jwt import current_identity, jwt_required
-from flask_security.utils import encrypt_password
 from flask_security.recoverable import reset_password_token_status
 
 from flask_apispec import MethodResource, marshal_with, use_kwargs, doc
-from marshmallow import Schema, fields, validates, ValidationError, post_load
 from ..models.auth import User
 from ..database import db
 from ..auth import register_user, reset_password, send_confirmation
 from .utils import abort, auth_required
 from ..utils.db import put_record
-
-
-class UserSchema(Schema):
-    email = fields.Email(required=True)
-    name = fields.Str(required=True, description='User full name')
-    password = fields.Str(load_only=True,
-                          description='Password. Minimum 6 characters.')
-    picture = fields.Str(allow_none=True)
-    analyses = fields.Nested('AnalysisSchema',
-                             only=['hash_id', 'name', 'status',
-                                   'description', 'modified_at', 'dataset_id'],
-                             many=True, dump_only=True)
-
-    @validates('password')
-    def validate_pass(self, value):
-        if len(value) < 6:
-            raise ValidationError('Password must be at least 6 characters.')
-
-    @post_load
-    def encrypt_password(self, in_data):
-        if 'password' in in_data:
-            in_data['password'] = encrypt_password(in_data['password'])
-        return in_data
-
-    class Meta:
-        strict = True
-
-
-class UserCreationSchema(UserSchema):
-    password = fields.Str(load_only=True, required=True,
-                          description='Password. Minimum 6 characters.')
-
-    @validates('email')
-    def validate_name(self, value):
-        if User.query.filter_by(email=value).first():
-            raise ValidationError('Email already in use.')
-
-
-class UserResetSchema(UserCreationSchema):
-    token = fields.Str(required=True, description="Password reset token.")
+from ..schemas.user import UserSchema, UserCreationSchema, UserResetSchema
 
 # @doc(tags=['auth'])
 @marshal_with(UserSchema)
