+++ neuroscout/schemas/run.py
@@ -1,0 +1,13 @@
+from marshmallow import fields, Schema
+
+
+class RunSchema(Schema):
+    id = fields.Int()
+    session = fields.Str(description='Session number')
+    acquisition = fields.Str(description='Acquisition')
+    subject = fields.Str(description='Subject id')
+    number = fields.Int(description='Run id')
+    duration = fields.Number(description='Total run duration in seconds.')
+    dataset_id = fields.Int(description='Dataset run belongs to.')
+    task = fields.Nested(
+        'TaskSchema', only='id', description="Task id and name")
