+++ neuroscout/utils/core.py
@@ -3,6 +3,7 @@
 """
 from .. import resources
 
+
 def route_factory(app, docs, pairings, prepend='/api/'):
     """ Create API routes and add to app """
     for res_name, route in pairings:
@@ -10,6 +11,7 @@ def route_factory(app, docs, pairings, prepend='/api/'):
         app.add_url_rule(prepend + route,
                          view_func=res.as_view(res_name.lower()))
         docs.register(res)
+
 
 def listify(obj):
     ''' Wraps all non-list or tuple objects in a list; provides a simple way
